rename '_typing.py' to 'typing.py'
